requirements.txt
Django==3.2.12
djangorestframework==3.13.1
PyJWT==2.3.0
django-heroku==0.3.1